Initial Commit
Initially committing files to github for use on other devices. Experiencing issues committing from the command line but that should be worked out eventually. For now this will have to do. Next step is to add functionality to output mutiple waves through separate channels. This can easily be inferred through Calvin's code. From there the next step will be digital output and adjusting the display functionality of what is going on.