Memory Issue Fixed, Digital Output Not Working
Memory issue has been fixed. The issue was an accidental squaring and using radians as opposed to Hz. This issue has been resolved. However, once this issue was resolved and after working on simplifying the code for Digital Output, the digital functionality is not reading. One of two things is happening here. That is that either the digital output is not being generated properly in a list, or the amplitude is not high enough to output. Will explore this at a later date.